--- food-api/src/main/java/br/com/alfac/food/api/pedido/dto/ItemRequest.java
@@ -1,17 +1,0 @@
-package br.com.alfac.food.api.pedido.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-public class ItemRequest {
-    @Schema(example = "15")
-    private Long id;
-    
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-}
