--- food-api/src/main/java/br/com/alfac/food/api/pedido/dto/PedidosResponse.java
@@ -1,9 +1,0 @@
-package br.com.alfac.food.api.pedido.dto;
-
-import br.com.alfac.food.core.application.pedido.dto.PedidoDTO;
-
-import java.util.List;
-
-public record PedidosResponse (List<PedidoDTO> pedidos) {
-
-}
