--- food-api/src/main/java/br/com/alfac/food/api/pedido/mapper/PedidoMapper.java
@@ -1,13 +1,0 @@
-package br.com.alfac.food.api.pedido.mapper;
-
-import br.com.alfac.food.api.pedido.dto.PedidoRequest;
-import br.com.alfac.food.core.application.pedido.dto.PedidoDTO;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface PedidoMapper {
-
-    PedidoDTO toDTO(PedidoRequest pedidoRequest);
-
-}
