--- food-api/src/main/java/br/com/alfac/food/api/webhook/dto/RetornoPagamentoRequest.java
@@ -1,4 +1,0 @@
-package br.com.alfac.food.api.webhook.dto;
-
-public record RetornoPagamentoRequest(Long pagamentoId) {
-}
