+++ food-infra/src/main/java/br/com/alfac/food/infra/pagamento/webhook/dto/RetornoPagamentoRequest.java
@@ -1,0 +1,4 @@
+package br.com.alfac.food.infra.pagamento.webhook.dto;
+
+public record RetornoPagamentoRequest(Long pagamentoId) {
+}
