+++ pom.xml
@@ -9,7 +9,7 @@
         <module>food-client</module>
         <module>food-core</module>
         <module>food-database</module>
-        <module>food-api</module>
+        <module>food-infra</module>
     </modules>
     <parent>
         <groupId>org.springframework.boot</groupId>
